misc warning elimination

+++ Main.cpp
@@ -23,7 +23,7 @@ int main(int argc, char *argv[])
 
         scene = new Scene(xmlPath);
 
-        for (int i = 0; i < scene->cameras.size(); i++)
+        for (size_t i = 0; i < scene->cameras.size(); i++)
         {
             // initialize image with basic values
             scene->initializeImage(scene->cameras[i]);
